lab 5

+++ Problem#3/stepCounter.py
@@ -1,9 +1,9 @@
 def feet_to_steps(user_feet):
-   return(user_feet/2.5)
+   return(int(user_feet/2.5))
 
 if __name__ == '__main__':
     #take input feet steps here
     #store it into the function
     user_feet = float(input()) 
     #print your steps here
-    print(int(feet_to_steps(user_feet)))
+    print(feet_to_steps(user_feet))
